Organized some library files and finished recorder
DOESN'T WORK RN. Need to go through Joycon_rs docs to try and figure out
how to actually write my own recorder implementation that isn't such a
mess that's difficult to understand. Goal right now is to print a data
set to the screen after recording it from a joycon.